# feat(pooling): faster avg. with EmbeddingBags