@fdschmidt93 2024-10-03
Given the importance of the function, please briefly checkout the first commit and run the above function to verify you get the same output @vaibhavad :sweat_smile: :)